requirements.txt
streamlit>=1.25.0
pandas>=1.5.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
plotly>=5.15.0
Pillow>=9.4.0
